Create split_testing.md
Use Split testing or [A/B testing] (https://sendgrid.com/blog/sendgrid-email-marketing-service-tutorial-how-to-ab-split-test/) to test how different factors affect the way subscribers engage with campaigns. A/B testing allows 
send two different versions of your email campaign to randomly selected parts of your list. The version with the
most opens or clicks is then automatically sent to the rest of your list.

A/B Split Campaigns are an advanced feature available on the free plan for testing and on silver plans and higher 
through the [Marketing Application] (https://sendgrid.com/docs/Marketing_Emails/index.html) or 
[Marketing api] (https://sendgrid.com/docs/API_Reference/Marketing_Emails_API/index.html).